feat: perspective card bookmark changes
