+++ blocks/bookmarks/bookmarks.js
@@ -6,11 +6,11 @@ import { fetchArticleByID } from '../../scripts/data-service/article-data-servic
 import { CONTENT_TYPES } from '../../scripts/browse-card/browse-cards-constants.js';
 import BuildPlaceholder from '../../scripts/browse-card/browse-card-placeholder.js';
 import { bookmarksEventEmitter } from '../../scripts/events.js';
+import { getCardData, convertToTitleCase } from '../../scripts/utils/article-utils.js';
 import Pagination from '../../scripts/pagination/pagination.js';
 
 const BOOKMARKS_BY_PG_CONFIG = {};
 const CARDS_MODEL = {};
-const convertToTitleCase = (str) => (str ? str.replace(/\b\w/g, (match) => match.toUpperCase()) : '');
 
 const buildCardsShimmer = new BuildPlaceholder(Pagination.getItemsCount());
 
@@ -23,7 +23,14 @@ try {
 }
 
 export const parse = (model) => {
-  const { id, Type = [], Thumbnail, Solution = [], Title, URL, Description, Role = [] } = model;
+  const { id, Type = [], Thumbnail, Solution = [], Title, URL, Description, Role = [], viewLinkText } = model;
+  if (viewLinkText) {
+    // Already parsed as part of getCardData. No need to parse again.
+    if (!model.badgeTitle) {
+      model.badgeTitle = model.contentType;
+    }
+    return model;
+  }
   const [contentType = 'Course'] = Type;
   const [role] = Role;
   const contentTypeTitleCase = convertToTitleCase(contentType?.toLowerCase());
@@ -53,11 +60,17 @@ async function renderCards({ pgNum, block }) {
   buildCardsShimmer.add(wrapper.parentElement);
   wrapper.style.display = 'none';
   const bookmarkIds = BOOKMARKS_BY_PG_CONFIG[pgNum];
-  const bookmarkPromises = bookmarkIds.map((bookmarkId) => fetchArticleByID(bookmarkId));
+  const bookmarkPromises = bookmarkIds.map((bookmarkId) => {
+    if (bookmarkId.startsWith('/')) {
+      return getCardData(`${window.hlx.codeBasePath}${bookmarkId}`, placeholders);
+    }
+    return fetchArticleByID(bookmarkId);
+  });
   const cardResponse = await Promise.all(bookmarkPromises);
+
   const cardsData = cardResponse.filter(Boolean).map((card) => {
     const parsedCard = parse(card);
-    if (!CARDS_MODEL[parsedCard.id]) {
+    if (parsedCard.id && !CARDS_MODEL[parsedCard.id]) {
       CARDS_MODEL[parsedCard.id] = parsedCard;
     }
     return parsedCard;
